Fixed failing Stripe webhook test
refs https://github.com/TryGhost/Ghost/commit/477c3c73260bdf5fc0450a5caa6affd5e998bb85

Due to the bump in express-test, we now handle string bodies 'properly'. So they now pass all the Express middlewares. In the past this failing test did not really pass by the bodyParser.raw middleware,
so the content-type check on the `bodyParser.raw({type: 'application/json'})` middleware was not executed. Now it is, and the test fails because the content-type header was not set to application/json.