Moved assets to AdminX Design System
refs. https://github.com/TryGhost/Team/issues/3318

Until now all assets was part of the main app and not the AdminX Design System which could be problematic later when the system becomes a standalone library. Moved the assets folder under the design system for better reusability, less redundancy and for a more self-encapsulated system